+++ docs/version_history.md
@@ -31,3 +31,4 @@
 * Make "List Max" buttons display a warning when tutorial is not maxed
 * Make "List Max" buttons max the tutorial when tutorial is not maxed
 * Remove Dream Summon vouchers from Summon Ticket list
+* Fix import/export duplicate weapon_passive_id's (1.3.1)